requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pgvector
google-genai
PyMuPDF
python-dotenv
python-multipart
